+++ xbmc/utils/RingBuffer.cpp
@@ -31,7 +31,7 @@ CRingBuffer::~CRingBuffer()
 }
 
 /* Create a ring buffer with the specified 'size' */
-bool CRingBuffer::Create(unsigned int size)
+bool CRingBuffer::Create(size_t size)
 {
   CSingleLock lock(m_critSection);
   m_buffer = (char*)malloc(size);
@@ -70,7 +70,7 @@ void CRingBuffer::Clear()
 /* Read in data from the ring buffer to the supplied buffer 'buf'. The amount
  * read in is specified by 'size'.
  */
-bool CRingBuffer::ReadData(char* buf, unsigned int size)
+bool CRingBuffer::ReadData(char* buf, size_t size)
 {
   CSingleLock lock(m_critSection);
   if (size > m_fillCount)
@@ -79,7 +79,7 @@ bool CRingBuffer::ReadData(char* buf, unsigned int size)
   }
   if (size + m_readPtr > m_size)
   {
-    unsigned int chunk = m_size - m_readPtr;
+    size_t chunk = m_size - m_readPtr;
     memcpy(buf, m_buffer + m_readPtr, chunk);
     memcpy(buf + chunk, m_buffer, size - chunk);
     m_readPtr = size - chunk;
@@ -98,7 +98,7 @@ bool CRingBuffer::ReadData(char* buf, unsigned int size)
 /* Read in data from the ring buffer to another ring buffer object specified by
  * 'rBuf'.
  */
-bool CRingBuffer::ReadData(CRingBuffer& rBuf, unsigned int size)
+bool CRingBuffer::ReadData(CRingBuffer& rBuf, size_t size)
 {
   CSingleLock lock(m_critSection);
   if (rBuf.getBuffer() == NULL)
@@ -107,7 +107,7 @@ bool CRingBuffer::ReadData(CRingBuffer& rBuf, unsigned int size)
   bool bOk = size <= rBuf.getMaxWriteSize() && size <= getMaxReadSize();
   if (bOk)
   {
-    unsigned int chunksize = std::min(size, m_size - m_readPtr);
+    size_t chunksize = std::min(size, m_size - m_readPtr);
     bOk = rBuf.WriteData(&getBuffer()[m_readPtr], chunksize);
     if (bOk && chunksize < size)
       bOk = rBuf.WriteData(&getBuffer()[0], size - chunksize);
@@ -121,7 +121,7 @@ bool CRingBuffer::ReadData(CRingBuffer& rBuf, unsigned int size)
 /* Write data to ring buffer from buffer specified in 'buf'. Amount read in is
  * specified by 'size'.
  */
-bool CRingBuffer::WriteData(const char* buf, unsigned int size)
+bool CRingBuffer::WriteData(const char* buf, size_t size)
 {
   CSingleLock lock(m_critSection);
   if (size > m_size - m_fillCount)
@@ -130,7 +130,7 @@ bool CRingBuffer::WriteData(const char* buf, unsigned int size)
   }
   if (size + m_writePtr > m_size)
   {
-    unsigned int chunk = m_size - m_writePtr;
+    size_t chunk = m_size - m_writePtr;
     memcpy(m_buffer + m_writePtr, buf, chunk);
     memcpy(m_buffer, buf + chunk, size - chunk);
     m_writePtr = size - chunk;
@@ -149,7 +149,7 @@ bool CRingBuffer::WriteData(const char* buf, unsigned int size)
 /* Write data to ring buffer from another ring buffer object specified by
  * 'rBuf'.
  */
-bool CRingBuffer::WriteData(CRingBuffer& rBuf, unsigned int size)
+bool CRingBuffer::WriteData(CRingBuffer& rBuf, size_t size)
 {
   CSingleLock lock(m_critSection);
   if (m_buffer == NULL)
@@ -158,8 +158,8 @@ bool CRingBuffer::WriteData(CRingBuffer& rBuf, unsigned int size)
   bool bOk = size <= rBuf.getMaxReadSize() && size <= getMaxWriteSize();
   if (bOk)
   {
-    unsigned int readpos = rBuf.getReadPtr();
-    unsigned int chunksize = std::min(size, rBuf.getSize() - readpos);
+    size_t readpos = rBuf.getReadPtr();
+    size_t chunksize = std::min(size, rBuf.getSize() - readpos);
     bOk = WriteData(&rBuf.getBuffer()[readpos], chunksize);
     if (bOk && chunksize < size)
       bOk = WriteData(&rBuf.getBuffer()[0], size - chunksize);
@@ -177,14 +177,14 @@ bool CRingBuffer::SkipBytes(int skipSize)
     return false; // skipping backwards is not supported
   }
 
-  unsigned int size = skipSize;
+  size_t size = skipSize;
   if (size > m_fillCount)
   {
     return false;
   }
   if (size + m_readPtr > m_size)
   {
-    unsigned int chunk = m_size - m_readPtr;
+    size_t chunk = m_size - m_readPtr;
     m_readPtr = size - chunk;
   }
   else
@@ -216,30 +216,30 @@ char* CRingBuffer::getBuffer()
   return m_buffer;
 }
 
-unsigned int CRingBuffer::getSize()
+size_t CRingBuffer::getSize()
 {
   CSingleLock lock(m_critSection);
   return m_size;
 }
 
-unsigned int CRingBuffer::getReadPtr() const
+size_t CRingBuffer::getReadPtr() const
 {
   return m_readPtr;
 }
 
-unsigned int CRingBuffer::getWritePtr()
+size_t CRingBuffer::getWritePtr()
 {
   CSingleLock lock(m_critSection);
   return m_writePtr;
 }
 
-unsigned int CRingBuffer::getMaxReadSize()
+size_t CRingBuffer::getMaxReadSize()
 {
   CSingleLock lock(m_critSection);
   return m_fillCount;
 }
 
-unsigned int CRingBuffer::getMaxWriteSize()
+size_t CRingBuffer::getMaxWriteSize()
 {
   CSingleLock lock(m_critSection);
   return m_size - m_fillCount;
